Revert "moved all subprograms to a new subprogram folder, to clean up the main folder"
This reverts commit e734547df9de85582db5f1cf418fbbf32244faca.
