+++ aufbau.py
