+++ ausgabe.py
